msvc snprintf fix; image/image.c double->float compiler warnings remvoed
git-svn-id: http://svn.xvid.org/branches/dev-api-3@810 060a26f8-e6a2-0410-8959-474133d486e0

+++ xvidcore/src/image/image.c
@@ -942,9 +942,9 @@ image_psnr(IMAGE * orig_image,
 float sse_to_PSNR(long sse, int pixels)
 {
         if (sse==0)
-                return 99.99;
+                return 99.99F;
 
-        return 48.131 - 10*log10((float)sse/(float)(pixels));   // log10(255*255)=4.8131
+        return 48.131F - 10*(float)log10((float)sse/(float)(pixels));   // log10(255*255)=4.8131
 
 }
